Refactor

+++ dataset.py
@@ -104,7 +104,7 @@ def rescale_image(image, scale):
     size = tf.shape(image)[:2]
     shorter = tf.argmin(size)
     ratio = scale / size[shorter]
-    new_size = tf.to_int32(tf.round(size * ratio))
+    new_size = tf.to_int32(tf.round(tf.to_float(size) * ratio))
 
     return tf.image.resize_images(image, new_size, method=tf.image.ResizeMethod.BILINEAR)
